refactor(Jupyter kernels): Factor out `kernel` module
